Cambio a double Precio

+++ src/domain/Videojuego.java
@@ -10,7 +10,7 @@ public class Videojuego extends Producto{
 	private TipoConsola tipo;
 	private String autor;
 	
-	public Videojuego(long id, String titulo, String sinopsis, float precio, int rating, ArrayList<Review> comentarios,
+	public Videojuego(long id, String titulo, String sinopsis, double precio, int rating, ArrayList<Review> comentarios,
 			GeneroVideoJuego genero, TipoConsola tipo, String autor, ImageIcon foto) {
 		super(id,titulo, sinopsis, precio, rating, comentarios, foto);
 		this.genero = genero;
